Put Python docs attributes in a table, which includes linking for more info
As part of this we overload a couple of templates in order to customise
them. We remove a trailing colon (sounds like localisation is not fully
implemented, so changing a dictionary is not an option), and remove
the default place where attributes are listed as we now display them
in the summary.